fix: refactor javascript code structure
