+++ lib/inspect.js
@@ -1,12 +1,13 @@
 
 const chalk = require('chalk')
 
-const nodegtk = require('../lib/index');
-nodegtk.startLoop();
+const internal = require('./native.js')
+const gi = require('../lib/index');
+gi.startLoop();
 
 const infos = []
 
-const GI = nodegtk._GIRepository;
+const GI = gi._GIRepository;
 
 function def(obj, name, data) {
     Object.defineProperty(obj, name, {
@@ -152,7 +153,7 @@ function TypeInfo(info) {
         this.array_type = GI.ArrayType[GI.type_info_get_array_type(info)];
         this.zero_terminated = GI.type_info_is_zero_terminated(info);
         this.fixed_size = GI.type_info_get_array_fixed_size(info);
-        this.size = nodegtk._c.GetTypeSize(info);
+        this.size = internal.GetTypeSize(info);
         const isPointer = GI.type_info_is_pointer(info)
         if (isPointer)
             this.isPointer = isPointer
@@ -188,7 +189,7 @@ function TypeInfo(info) {
     }
     else {
         this.type = infoTypeString(info);
-        this.size = nodegtk._c.GetTypeSize(info);
+        this.size = internal.GetTypeSize(info);
         const isPointer = GI.type_info_is_pointer(info)
         if (isPointer)
             this.isPointer = isPointer
@@ -197,7 +198,7 @@ function TypeInfo(info) {
 
 function ConstantInfo(info, parent) {
     BaseInfo.call(this, info, parent)
-    this.value = nodegtk._c.GetConstantValue(info);
+    this.value = internal.GetConstantValue(info);
 }
 
 function ValueInfo (info, parent) {
@@ -292,7 +293,7 @@ function UnionInfo(info, parent) {
         this.fields[fieldName] = new FieldInfo(field, this);
         if (is_discriminated) {
             const d = GI.union_info_get_discriminator(info, i);
-            this.fields[fieldName].discriminator = nodegtk._c.GetConstantValue(d);
+            this.fields[fieldName].discriminator = internal.GetConstantValue(d);
         }
     }
 }
@@ -534,7 +535,7 @@ function parseNamespace(ns, ver) {
     const module = Object.create(null);
     const repo = GI.Repository_get_default();
 
-    if (!nodegtk._isLoaded(ns, ver))
+    if (!gi._isLoaded(ns, ver))
         GI.Repository_require.call(repo, ns, ver, 0);
 
     const nInfos = GI.Repository_get_n_infos.call(repo, ns);
