+++ lib/overrides/GObject.js
@@ -3,6 +3,7 @@
  */
 
 const gi = require('..')
+const internal = require('./native.js')
 const system = gi.System
 
 
@@ -40,11 +41,11 @@ exports.apply = (GObject) => {
 
   GObject.Object.prototype._setProperty = GObject.Object.prototype.setProperty
   GObject.Object.prototype.setProperty = function (name, value) {
-    gi._c.ObjectPropertySetter(this, name, value)
+    internal.ObjectPropertySetter(this, name, value)
   }
   GObject.Object.prototype._getProperty = GObject.Object.prototype.getProperty
   GObject.Object.prototype.getProperty = function (name) {
-    return gi._c.ObjectPropertyGetter(this, name)
+    return internal.ObjectPropertyGetter(this, name)
   }
   
   GObject.Value.prototype.getBoxed = function getBoxed() {
